Project ARGUS-V2/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "argus-v2"
version = "2.0.0"
description = "High-Performance Multi-Agent AI Orchestration Framework"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "ARGUS Team", email = "team@argus-framework.dev"},
]
keywords = ["ai", "orchestration", "multi-agent", "async", "claude", "openai", "gemini"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "aiohttp>=3.9.0",
    "structlog>=23.2.0",
    "typer[all]>=0.9.0",
    "rich>=13.7.0",
    "psutil>=5.9.0",
    "pyyaml>=6.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-benchmark>=4.0.0",
    "ruff>=0.1.6",
    "mypy>=1.7.0",
    "pre-commit>=3.5.0",
    "bandit>=1.7.5",
]

[project.urls]
Homepage = "https://github.com/your-username/argus-v2"
Repository = "https://github.com/your-username/argus-v2"
Issues = "https://github.com/your-username/argus-v2/issues"

[project.scripts]
argus = "argus_core.cli:app"

[tool.setuptools.packages.find]
include = ["argus_core*"]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]

[tool.ruff]
line-length = 88
target-version = "py311"